+++ Razor/UI/Razor.cs
@@ -523,6 +523,8 @@ namespace Assistant
 
             nextPrevAbcOrder.SafeAction(s => { s.Checked = Config.GetBool("NextPrevAlphabetical"); });
 
+            playEmoteSound.SafeAction(s => { s.Checked = Config.GetBool("PlayEmoteSound"); });
+
             Engine.MainWindow.Size = new Size(Config.GetInt("WindowSizeX"), Config.GetInt("WindowSizeY"));
 
             // Disable SmartCPU in case it was enabled before the feature was removed
@@ -8126,6 +8128,11 @@ namespace Assistant
             }
 
             return filteredNodes.ToArray();
+        }
+
+        private void playEmoteSound_CheckedChanged(object sender, EventArgs e)
+        {
+            Config.SetProperty("PlayEmoteSound", playEmoteSound.Checked);
         }
     }
 }